backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-genai==1.14.0
python-dotenv==1.1.0
pillow==11.1.0
pydantic==2.10.0
python-multipart==0.0.18
httpx==0.28.0
gunicorn==23.0.0
